pyproject.toml
[project]
name = 'YaMDb'
version = '0.1.1'
description = 'The YaMDb project collects users feedback on various content'
readme = 'README.md'
requires-python = '==3.9.13'
dependencies = [
    'requests==2.26.0',
    'Django==3.2',
    'djangorestframework==3.12.4',
    'PyJWT==2.1.0',
    'pytest==6.2.4',
    'pytest-django==4.4.0',
    'pytest-pythonpath==0.7.3',
    'django-filter==23.5',
    "djangorestframework-simplejwt>=5.3.1",
    "pandas>=2.2.3",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
exclude = ["postman_collection*"]

[tool.ruff]
target-version = 'py39'
line-length = 79
fix = true

[tool.ruff.format]
quote-style = 'single'
docstring-code-format = true

[tool.isort]
line_length = 79
multi_line_output = 3
use_parentheses = true
combine_as_imports = true
skip_gitignore = true
skip = ["migrations", "tests"]
include_trailing_comma = true
known_local_folder = ['api', 'reviews', 'api_yamdb']
known_third_party = ["django", "rest_framework"]
sections = ["STDLIB", "THIRDPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
